update schemas to include JSON Schema draft-07 and add required properties; refactor test paths and add run-tests script
