+++ schemas/limit.json
@@ -1,7 +1,7 @@
 {
-  "type": "object",
+  "$schema": "http://json-schema.org/draft-07/schema#",
   "$id": "schema:limit",
-  "additionalProperties": false,
+  "type": "object",
   "properties": {
     "limit": {
       "type": "integer",
