+++ schemas/loader.js
@@ -1,8 +1,8 @@
-const fp = require("fastify-plugin");
-module.exports = fp(
-  function (fastify, opts, next) {
-    fastify.addSchema(require("./dotenv.json"));
-    next();
-  },
-  { name: "application-schemas" },
-);
+// const fp = require("fastify-plugin");
+// module.exports = fp(
+//   function (fastify, opts, next) {
+//     fastify.addSchema(require("./dotenv.json"));
+//     next();
+//   },
+//   { name: "application-schemas" },
+// );
