+++ schemas/skip.json
@@ -1,6 +1,7 @@
 {
-  "type": "object",
+  "$schema": "http://json-schema.org/draft-07/schema#",
   "$id": "schema:skip",
+  "type": "object",
   "additionalProperties": false,
   "properties": {
     "skip": {
